+++ src/main/java/com/andreidodu/europealibrary/batch/indexer/step/metainfo/dataextractor/strategy/PdfMetaInfoExtractorStrategyImpl.java
@@ -1,7 +1,6 @@
 package com.andreidodu.europealibrary.batch.indexer.step.metainfo.dataextractor.strategy;
 
 import com.andreidodu.europealibrary.batch.indexer.step.common.StepUtil;
-import com.andreidodu.europealibrary.batch.indexer.step.metainfo.TmpAssociationUtil;
 import com.andreidodu.europealibrary.batch.indexer.step.metainfo.dataextractor.MetaInfoExtractorStrategy;
 import com.andreidodu.europealibrary.constants.DataPropertiesConst;
 import com.andreidodu.europealibrary.dto.BookCodesDTO;
@@ -10,6 +9,7 @@ import com.andreidodu.europealibrary.model.FileMetaInfo;
 import com.andreidodu.europealibrary.model.FileSystemItem;
 import com.andreidodu.europealibrary.repository.BookInfoRepository;
 import com.andreidodu.europealibrary.repository.FileMetaInfoRepository;
+import com.andreidodu.europealibrary.service.TmpAssociationService;
 import com.andreidodu.europealibrary.util.FileUtil;
 import com.andreidodu.europealibrary.util.PdfUtil;
 import com.andreidodu.europealibrary.util.StringUtil;
@@ -43,7 +43,7 @@ public class PdfMetaInfoExtractorStrategyImpl implements MetaInfoExtractorStrate
     private final OtherMetaInfoExtractorStrategyImpl otherMetaInfoExtractorStrategy;
     private final FileUtil fileUtil;
     private final StepUtil stepUtil;
-    private final TmpAssociationUtil tmpAssociationUtil;
+    private final TmpAssociationService tmpAssociationService;
     @Value("${com.andreidodu.europea-library.job.indexer.step-indexer.disable-pdf-metadata-extractor}")
     private boolean disablePDFMetadataExtractor;
     @Value("${com.andreidodu.europea-library.job.indexer.step-meta-info-writer.disable-isbn-extractor}")
@@ -103,7 +103,7 @@ public class PdfMetaInfoExtractorStrategyImpl implements MetaInfoExtractorStrate
             final String keywordsStringTrimmed = StringUtil.cleanAndTrimToNull(documentInformation.getKeywords());
             List<String> keywords = new ArrayList<>();
             keywords.add(keywordsStringTrimmed);
-            this.tmpAssociationUtil.addItemsToTmpAssociationTable(savedFileMetaInfo.getId(), keywords);
+            this.tmpAssociationService.addItemsToTmpAssociationTable(savedFileMetaInfo.getId(), keywords);
         } catch (Exception e) {
             log.debug("invalid pdf book keywords for '{}'", fullPath);
         }
