+++ src/main/java/com/andreidodu/europealibrary/service/impl/TmpAssociationServiceImpl.java
@@ -1,9 +1,10 @@
-package com.andreidodu.europealibrary.batch.indexer.step.metainfo;
+package com.andreidodu.europealibrary.service.impl;
 
 import com.andreidodu.europealibrary.batch.indexer.step.common.StepUtil;
 import com.andreidodu.europealibrary.constants.DataPropertiesConst;
 import com.andreidodu.europealibrary.model.TmpAssociation;
 import com.andreidodu.europealibrary.repository.TmpAssociationRepository;
+import com.andreidodu.europealibrary.service.TmpAssociationService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Propagation;
@@ -13,10 +14,11 @@ import java.util.List;
 
 @Component
 @RequiredArgsConstructor
-public class TmpAssociationUtil {
+public class TmpAssociationServiceImpl implements TmpAssociationService {
     private final TmpAssociationRepository tmpAssociationRepository;
     private final StepUtil stepUtil;
 
+    @Override
     @Transactional(propagation = Propagation.REQUIRES_NEW)
     public void addItemsToTmpAssociationTable(Long targetId, List<String> items) {
         this.tmpAssociationRepository.saveAllAndFlush(
